requirements.txt
# Deep Learning Framework
torch>=1.10.0
torchvision>=0.11.0
timm>=0.6.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
astropy>=5.0.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0

# Machine Learning Metrics
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Model Analysis
thop>=0.1.0
torchsummary>=1.5.0

# Data Download
wget>=3.2

# Utilities
tqdm>=4.62.0
